Make configuration more flexible
- **Add environment variable for config location**: in case the default './config.json' isn't suitable for some deployment, it's nice to have the option to change it
- **Allow secrets to be defined through environment variables**: I don't love that I have to store my DSN and password in the config file, and that it's logged on startup. I'd much rather those values be configurable through the environment

I also moved all the configuration-related logic into its own file, since it is kind of scattered between main.go and utils.go, and added some small formatting fixes, courtesy of `go fmt`.
